modified:   CMakeLists.txt     modified:   README.md     modified:   src/WallpaperEngine/Render/Drivers/Detectors/KDEWaylandFullScreenDetector.cpp     new file:   src/WallpaperEngine/Render/Drivers/Detectors/KDEWaylandFullScreenDetector.h     deleted:    src/WallpaperEngine/Render/Drivers/Detectors/KDEWaylandFullScreenDetector.hpp     modified:   src/WallpaperEngine/Render/Drivers/Detectors/WaylandFullScreenDetector.cpp Renamed KDEWaylandFullScreenDetector.hpp to KDEWaylandFullScreenDetector.h to better match naming conventions of this project. Cleaned up the fallback logic in KDEWaylandFullScreenDetector to avoid unnecessary complexity and potential issues with the fallback detector. Made DBUS libs no longer a hard dependency, instead they are only used if the project is compiled with the ENABLE_KDE_WAYLAND_FULLSCREEN_DETECTOR flag; since DBUS is a required dependency of KDE Plasma no need to update the README.md to mention the need for this dependency as systems running KDE Plasma will already have it installed. Updated README.md to add instructions for enabling the KDE Wayland fullscreen detection feature, mention of the need for a special script is now also in the same place so it's easier to find, replaced [here] with descriptive text as the markdown link target to improve accessibility and satisfy markdownlint.
