+++ README.md
@@ -138,6 +138,22 @@ cmake -DCMAKE_BUILD_TYPE='Release' ..
 make
 ```
 
+> ⚠️ Note for KDE Plasma users:
+
+KDE Plasma (Wayland) users who intend to use the “pause on fullscreen” feature are required to install an additional script from the [KWin Maximize Detector](https://github.com/LS-FCEFyN/Maximize-Detector) repository. They must also replace:
+
+```bash
+cmake -DCMAKE_BUILD_TYPE='Release' ..
+```
+
+With:
+
+```bash
+cmake -DCMAKE_BUILD_TYPE='Release' -DENABLE_KDE_EXPERIMENTAL_FEATURES=ON ..
+```
+
+
+
 Once the build process is finished, this should create a new `output` folder containing the app and all the required
 support files to run.
 
@@ -297,7 +313,6 @@ linux-wallpaperengine --set-property bloom=1 2370927443
 ## 🧪 Wayland & X11 Support
 
 - **Wayland**: Works with compositors that support `wlr-layer-shell-unstable`.
-- **KDE Wayland**: Users of KDE Plasma under Wayland that may wish to make use of the "pause on fullscreen" features may do so but require the installation of an additional script found [here](https://github.com/LS-FCEFyN/Maximize-Detector)
 - **X11**: Requires XRandr. Use `--screen-root <screen_name>` (as shown in `xrandr`).
 
 > ⚠ For X11 users: Currently doesn't work if a compositor or desktop environment (e.g. GNOME, KDE, Nautilus) is drawing the background.
